requirements.txt
Flask~=2.3.2
pytest~=7.4.0

openai~=0.27.8
tiktoken~=0.3.3

python_dotenv
openai-whisper
ffmpeg-python

torch
python-dotenv
